condense_metabolite_set

this function condense the metabolic-enzyme set collections to summarise ignore specific types of information such as compartments of relative positions (sign)

Usage

condense_metabolite_set(
  reaction_set_list,
  condense_sign = T,
  condense_compartments = F,
  ignore_sign = F
)

Arguments

reaction_set_list

list, metabolic-enzyme set collections as returned by the prepare_metabolite_set function

condense_sign

boolean, summarise upstream and downstream position of metabolites into a signle relative position

condense_compartments

boolean, wether the compartment information of hte metabolic network sohuld be ignored

ignore_sign

boolean, wether the upstream/downstream position of metabolic should be ignored

Value

a metabolic-enzyme set collections